REF: Organize Fitter file structure.
