+++ lmfit/__init__.py
@@ -24,7 +24,7 @@ from .printfuncs import (fit_report, ci_report,
 from .wrap  import wrap_function, make_paras_and_func
 from .model import Model
 from . import models
-from .fitter import Fitter
+from .ui import Fitter
 
 from . import uncertainties
 from .uncertainties import ufloat, correlated_values
